feat(databases): 数据库管理域全量实现——GarnetDatabase/IDatabaseManager/基座/单库/多库/工厂
- garnet_database(新): 逻辑库容器（共享 WedbStore + 前缀隔离 + 设备句柄 + AOF + 保存点）
- i_database_manager: 24 方法 trait（AFIT→RPIT impl Future）
- database_manager_base: 检查点(wkv CPR FoldOver)/AOF 重放(自描述记录格式)/过期扫描(wkv 原生)/混合日志统计/快照保留清理
- single/multi: 单库直通；多库 papaya 注册表 + SWAPDB 内容锁 + 按库检查点目录
- factory: DatabaseManager 枚举静态分发
- 依赖: papaya(并发注册表), waof(AOF), compio/tempfile(dev, 测试用)

+++ node/wserver/src/storage/session/objectstore/list_ops.rs
@@ -222,7 +222,7 @@ impl<'a, D: Device> StorageSession<'a, D> {
         let (start, step, skip): (i64, i64, u64) = if rank > 0 {
           (0, 1, (rank - 1) as u64)
         } else {
-          (guard.len() as i64 - 1, -1, (rank.unsigned_abs() - 1) as u64)
+          (guard.len() as i64 - 1, -1, rank.unsigned_abs() - 1)
         };
         let mut hits = 0u64;
         let mut i = start;
